fix: separate template overview and raw HTML route to show proper preview page
