+++ src/components/builder/WebsitePreview.tsx
@@ -65,7 +65,9 @@ export function WebsitePreview() {
       setIsLoading(true);
       setAssets(null);
       try {
-        const response = await fetch(`/api/templates/${encodeURIComponent(selectedTemplate.id)}`);
+        const response = await fetch(
+          `/api/templates/${encodeURIComponent(selectedTemplate.id)}?format=json`
+        );
         if (!response.ok) {
           throw new Error("Unable to load template");
         }
